Creación player.js
Con el fin de abstraer un poco y mejorar la semántica de los archivos
